+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.0.2
+[ToggleSheet] - Animate hiding of safe zone rectangle.
+[MultiStateSheet] - Animate hiding of safe zone rectangle 
+[MultiStateSheet] - Fix an issue where the outside widget would not display when the bottom sheet was opened with a height of 0px.
+[Package] - Improve docs.
+
 ## 1.0.1
 [ToggleSheet] - Fix bottom insets and safe area bottom padding rendering issues
 [MultiStateSheet] - Make footer `HitBehavior` default to `null` if its placed on top of content list